Move urlpatterns & views to pages app
