--- config/views.py
@@ -1,13 +1,0 @@
-from django.views.generic import TemplateView
-
-
-class HomePageView(TemplateView):
-    template_name = "home.html"
-
-
-class AboutPageView(TemplateView):
-    template_name = "about.html"
-
-
-class ProfilePageView(TemplateView):
-    template_name = "profile.html"
