+++ pages/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from .views import AboutPageView, HomePageView, ProfilePageView
+
+urlpatterns = [
+    path("", HomePageView.as_view(), name="home"),
+    path("about/", AboutPageView.as_view(), name="about"),
+    path("profile/", ProfilePageView.as_view(), name="profile"),
+]
