+++ pages/views.py
@@ -1,3 +1,13 @@
-from django.shortcuts import render  # noqa:F401
+from django.views.generic import TemplateView
 
-# Create your views here.
+
+class HomePageView(TemplateView):
+    template_name = "pages/home.html"
+
+
+class AboutPageView(TemplateView):
+    template_name = "pages/about.html"
+
+
+class ProfilePageView(TemplateView):
+    template_name = "pages/profile.html"
